+++ tests/configs/modules/calendar/diff_tz_start_end.js
@@ -17,6 +17,7 @@ let config = {
 				getRelative: 0,
 				maximumNumberOfDays: 28,
 				showEnd: true,
+				showEndsOnlyWithDuration: true,
 				calendars: [
 					{
 						maximumEntries: 100,
@@ -26,10 +27,6 @@ let config = {
 			}
 		}
 	]
-};
-
-Date.now = () => {
-	return new Date("08 Oct 2024 12:30:00 GMT-07:00").valueOf();
 };
 
 /*************** DO NOT EDIT THE LINE BELOW ***************/